libopencm3: Remove support
This and the CMSIS backend support exactly the same hardware and the
CMSIS backend is only a little more complicated but more interesting
to write and better optimized so that gets priority.

The backend is stable enough that adding this back shouldn't entail a
whole lot of work if it becomes desireable.